+++ src/main/java/org/cojen/maker/FieldMaker.java
@@ -26,6 +26,7 @@ public interface FieldMaker extends Maker {
     /**
      * Returns the name of this field.
      */
+    @Override
     String name();
 
     /**