+++ src/main/java/org/cojen/maker/TheClassMaker.java
@@ -136,6 +136,11 @@ final class TheClassMaker extends Attributed implements ClassMaker, Typed {
     }
 
     @Override
+    public String name() {
+        return type().name();
+    }
+
+    @Override
     public ClassMaker public_() {
         checkFinished();
         mModifiers = Modifiers.toPublic(mModifiers);
@@ -577,10 +582,6 @@ final class TheClassMaker extends Attributed implements ClassMaker, Typed {
         }
 
         return unimplemented == null ? Collections.emptySet() : unimplemented;
-    }
-
-    String name() {
-        return type().name();
     }
 
     void toModule() {