+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 v2.5.5
 ------
 * Added a feature to directly install classes into the class loader used by generated classes.
+* Added a method to obtain the ClassMaker name.
 
 v2.5.4 (2024-01-24)
 ------